Update export.rst
Import data in the galaxy library and in a new history
------------------------------------------------------
In the MetaGenSense side

Data to analyse, stored in the project directory (cf. 4th step), are copied within the user’s Galaxy library when he clicks on the “import new files” button. The files to analyze become datasets which will be copied within a galaxy history by clicking on the “create new history” button.